pyproject.toml
[tool.poetry]
name = "yml-resume"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pyyaml = "^6.0.1"
jinja2 = "^3.1.4"
playwright = "^1.44.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
